+++ tests/Checkout/Payment/PayPalPaymentHandlerTest.php
@@ -106,6 +106,30 @@ class PayPalPaymentHandlerTest extends TestCase
         $this->assertOrderTransactionState(OrderTransactionStates::STATE_UNCONFIRMED, $transactionId, Context::createDefaultContext());
     }
 
+    public function testPayAndFinalizePendingCapture(): void
+    {
+        $settings = $this->getDefaultConfigData();
+        $handler = $this->createPayPalPaymentHandler($settings);
+        $context = Context::createDefaultContext();
+
+        $transactionId = $this->getTransactionId($context, $this->getContainer());
+        $response = $handler->pay(new Request(), new PaymentTransactionStruct($transactionId), $context, null);
+        static::assertNotNull($response);
+
+        $this->assertOrderTransactionState(OrderTransactionStates::STATE_UNCONFIRMED, $transactionId, $context);
+
+        $this->orderTransactionRepo->update([[
+            'id' => $transactionId,
+            'customFields' => [
+                SwagPayPal::ORDER_TRANSACTION_CUSTOM_FIELDS_PAYPAL_ORDER_ID => CaptureOrderPending::ID,
+            ],
+        ]], $context);
+
+        $handler->finalize(new Request(), new PaymentTransactionStruct($transactionId), $context);
+
+        $this->assertOrderTransactionState(OrderTransactionStates::STATE_IN_PROGRESS, $transactionId, $context);
+    }
+
     public function testPayWithEcs(): void
     {
         $settings = $this->getDefaultConfigData();